Refactor HVAC temperature management and enablement logic
- Updated the HVAC class to replace the `enabled_mask` with `is_enabled_24h` for better handling of node enablement over a 24-hour period.
- Adjusted the temperature configuration to account for the new enablement structure.
- Modified the calculation of HVAC power output to utilize the new enablement logic.
- Cleaned up unnecessary code and comments related to the previous implementation.
- Ensured that the HVAC control logic respects the enabled state of nodes during optimization.

+++ python/digital_twin.py
@@ -68,7 +68,7 @@ class DigitalTwinService:
             end_ts = start_ts + pd.Timedelta(hours=24)
 
             self.simulation.current_time = start_ts
-            self.simulation.thermal_solver.T = np.array(list(cmd_json["t"].values()))
+            self.simulation.thermal_solver.T = np.array(list(cmd_json["t"].values()), dtype=float)
 
             self.simulation.hvac.set_temperatures_config()
             self.simulation.hvac.cached_plan = None
@@ -82,7 +82,6 @@ class DigitalTwinService:
     def run_physics_loop(self, end_timestamp):
         try:
             telemetry = []
-
             while self.simulation.current_time < end_timestamp:
                 step_data = self.simulation.run_step(self.simulation_step)
                 telemetry.append({"run_id": self.run_id, **step_data})
